Move endpoint functions into modules
